# src/config/docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TokenNode": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object"
			},
			"AuthResponse": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TokenNode"
							}
						],
						"nullable": true
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"default": 404
					},
					"messages": {
						"type": "string",
						"default": "Entity not found."
					},
					"data": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"InvalidCredentialsError": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"default": 409
					},
					"message": {
						"type": "string",
						"default": "Invalid credentials. Verify either your email or password."
					},
					"data": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"LoginParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-password_"
			},
			"AlreadyExistsError": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"default": 409
					},
					"message": {
						"type": "string",
						"default": "The resource presents conflicts, it's possible that some of the fields already exists or the modification is not possible."
					},
					"data": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.email-or-firstName-or-lastName-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"email",
					"password",
					"firstName",
					"lastName"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"SignUpParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-firstName-or-lastName-or-password_"
			},
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"surName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"middleName": {
						"type": "string",
						"nullable": true
					},
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"phoneCode": {
						"type": "string",
						"nullable": true
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"tokenVersion": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"username",
					"surName",
					"lastName",
					"middleName",
					"firstName",
					"phone",
					"phoneCode",
					"password",
					"email",
					"tokenVersion",
					"id"
				],
				"type": "object",
				"description": "Model User"
			},
			"UserResponseWithRecords": {
				"properties": {
					"data": {
						"properties": {
							"records": {
								"items": {
									"$ref": "#/components/schemas/User"
								},
								"type": "array"
							}
						},
						"required": [
							"records"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/User"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-firstName-or-lastName-or-password_"
			},
			"Pick_User.email-or-firstName-or-lastName-or-password-or-middleName-or-phone-or-phoneCode-or-surName-or-username_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"middleName": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"phoneCode": {
						"type": "string",
						"nullable": true
					},
					"surName": {
						"type": "string",
						"nullable": true
					},
					"username": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"email",
					"password",
					"firstName",
					"lastName",
					"middleName",
					"phone",
					"phoneCode",
					"surName",
					"username"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserUpdateParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-firstName-or-lastName-or-password-or-middleName-or-phone-or-phoneCode-or-surName-or-username_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ExTSPrimsaREST-BoilerPlate",
		"version": 1,
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "contact name, edit this in tsoa.json @ root.",
			"email": "contact email, edit this in tsoa.json @ root.",
			"url": "contact url, edit this in tsoa.json @ root."
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"description": "Authenticates with user credentials",
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginParams"
							}
						}
					}
				}
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Singup",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"description": "Creates an authenticates.\nSupply the required user fields and receive corresponding user details.",
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignUpParams"
							}
						}
					}
				}
			}
		},
		"/ping": {
			"get": {
				"operationId": "GetPing",
				"responses": {
					"200": {
						"description": "Success!",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Ping Controller"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseWithRecords"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of all users.",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"description": "Creates an user.\nSupply the required user fields and receive corresponding user details.",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing user.\nSupply the unique user ID from either and receive corresponding user details.",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"description": "Updates an user.\nSupply the required user fields and receive corresponding user details.",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:8000/v1/"
		}
	]
}